--- Json/Страница Авторизации.txt
@@ -1,36 +1,0 @@
-
-запрос-ответ json
-
-
-
-Запрос на авторизацию
-{
-	
-	'Table' : 'Users',
-	'action' : 'login',
-	'email' : string,
-	'pass' : string
-}
-
-Ответы:
-
-Положительный
-
-{
-         'id' : integer,
-         'name' : string,
-         'email' : string,
-         'fio' : string,
-         'avatar' : string,
-         'pass' : string,
-         'access' : string,
-         'subject' : string,                                 (только для учителей)
-         'class' : string                                   (только для учеников)
-}
-
-Отрицательный
-
-{
-	'result' : 'negative',
-	'message' : string
-}
