--- Json/Страница ЛК.txt
@@ -1,70 +1,0 @@
-
-Страница Личный Кабинет запрос-ответ json
-
-
-
-Запрос
-{
-	
-	'Table' : 'Users',
-	'action' : 'getUser',
-	'id' : integer,
-	'question' : string					
-}
-
-Ответы:
-
-Положительный
-
-{
-         'id' : integer,
-         'name' : string,
-         'email' : string,
-         'fio' : string,
-         'avatar' : string,
-         'pass' : string,
-         'access' : string,
-         'subject' : string,									       (только для учителей)
-         'class' : string											   (только для учеников)
-}
-
-Отрицательный
-
-{
-	'result' : 'negative',
-	'message' : string
-}
-
-------------------------------------------------------------
-
-Запрос на обновление полей 
-
-{
-	
-		 'Table' : 'Users',
-		 'action' : 'Update',
-		 'question' : string,
-		 'id' : integer,
-         ...
-         Ниже поля в соответсвии с названиями полей таблицы Users
-         только те поля, которые надо обновить пример:
-         'fio' : 'Иванов Иван Иваныч'
-         ...
-		 					
-}
-
-Ответы
-
-Положительный
-
-{
-	'result' : 'positive',
-	'message' : 'количетсво обновленных строк'
-}
-
-Отрицательный
-
-{
-	'result' : 'negative',
-	'message' : string
-}
