--- Json/Страница Оценок.txt
@@ -1,29 +1,0 @@
-
-Страница Журнал Оценок запрос-ответ json
-
-
-
-Запрос на все записи с оценками
-{
-	
-	"Table" : 'Answers',
-	"action" : 'scores',
-	'question' : string					
-}
-
-Ответы:
-
-Положительный
-
-{
-         'fio' : string,
-         'subject' : string,
-         'score' : integer
-}
-
-Отрицательный
-
-{
-	'result' : 'negative',
-	'message' : string
-}
