--- config/db.php
@@ -1,7 +1,0 @@
-<?php
-const DRIVER = 'mysql';
-const SERVER = 'localhost';
-const USERNAME = 'root';
-const PASSWORD = '';
-const DB = 'school';
-?>
